Add coder/configure example
Adds a directory configure/ containing example files for publishing
images that make use of Coder's configure feature. If an image contains
a script called "configure" in the directory "/coder", it is
automatically run during environment builds. The main use case is
configurations that depend on the $HOME directory that cannot otherwise
be added to an image's Dockerfile.